moved methods to save/load to Methods.py since they are very generic
